Refactor navbar layout styles
Menu content is now absolutely positioned instead so that it doesn't interrupt the flow of page content.

Button and H1 reworked to avoid the click block workaround by using flexbox instead. As the menu content is absolutely positioned it doesn't contribute to the nav dimensions, only the button does which plays well with flexbox here.

Decided to simplify the font sizing for H1.

Considered offsetting the page and menu content divs by the height of the navbar with a scss or css var, which is probably best if the navbar needs absolute or fixed positioning (although the menu content could continue to use the `top: 100%` approach), but have opted for now to just use the document flow.

Temporary debug colours are for better visualizing the sizing of layers (or nested element hierarchy) to get a better idea of the layout sizing going on, especially between mobile and desktop or other differences that affect scroll view sizing and behaviour.