Consistent scroll experience
Not exactly relevant to this project due to the actual view being too small, but these changes better handle the situation.

Notably 100vh for min-height prior would include the URL header bar of the Android Chrome UI, or on desktop Chrome if min-width was too small to trigger a scrollbar, it's height would be likewise included in the 100vh size, causing a vertical scrollbar unnecessarily.

A height of 100% avoids that issue on both mobile and desktop versions of Chrome, but would be dependent upon percentage of parent elements, which in turn could require the same of their parents, in this case `html > body > #root > .App`.

Originally I sized html to 100%, and could then use 100% on the body or 100vh with a max-height of 100% (though probably no point in 100vh then..?). `min-height: 100%` would work nicely at first, but child elements could not infer the height for their own `min-height: 100%` or `height: 100%` with regards to the parent if it was otherwise empty and not setting an explicit height?

Later realized it was just as effective to have a single parent layer use `position: absolute` which can size by 100% width and height of the viewport (on mobile Chrome that accounts for if the UI URL bar is present or not), then a child element can use `min-height: 100%` to size to that height or grow if it's content would be taller.

`min-height: 100%` is useful in a few scenarios where you'd want the element height to stretch to the bottom of the viewport, even if it's content is smaller. Often for layout purposes or stretching a bg colour or the layer itself.

`overflow-y: auto` was added to set the context for  which element/layer should be the scroll container. That affects the background-colour visually that would be seen, the behaviour is otherwise unaffected in this case.

`overflow-y` was not applied to the .App element as it's min-width would cause bad UX in the unlikely event  the viewport is smaller than the 180px limit that's been defined.