Minor layout adjustments
Shifted the padding to .App layer, as this encomposses both navbar and page components, min-width is set on .App and using min-width in nested components was causing the padding to not render as intended. This should work better.

---

The page component will now use the full height available, allowing it to center it's content without affecting the navbar position.

Considered removing navbar from a flex container, but currently the page component is dependent on sizing it's height correctly by being a flex item to use flex-grow.

Switching to an absolute positioned navbar (header element) should resolve this, and better support utilizing a transparent overlay over the page content. Quite possibly improving control over overflow too?