Refactor submit method + expiry validation fix
Fixes a bug with `isNotExpired()` check, where the month would be invalid for future years and past years would only be invalid depending on the month..

Would have been easier to just compare a timestamp of year and month together instead, but instead split the conditional check to identify/focus on the field value causing the error.

That was additionally required as the years `min(currentYear)` validation isn't valid anymore due to adding a String to Number transformation in the schema. This prevents doing validations on the primitive type, only transforms and refines are valid now; at least until zod adds the pre-process feature.

---

Split out the submission handling to `onValid()` and `onInvalid()` methods, although this required a new import to keep TS happy.

Due to zod schema changes we don't need to `parseInt` pre-process the input data, simplifying the submission logic.