Add aria attributes for inputs
Formatting hints for the month and year expiry date fields since they lack an equivalent label element. Seems better than the shared legend which is read out as MM YY?

If an error is present `aria-invalid` can be set to true which should also be helpful, although this doesn't seem to work on Android Talkback Chrome or Firefox in my testing.

I went with aria-describedby to additionally announce the error and it's related message.

Both use `undefined` if false to avoid rendering with the attributes set to "false".

---

The error list additionally is using `aria-live` and `aria-relevant` (albeit the values for that I believe are the expected defaults), but on Chrome the announcements aren't too useful as all values / elements are being announced in random order..