Improve aria attributes on the ErrorList component
`aria-atomic` is used here to ensure both the label and message are announced together.

Unfortunately with Chrome on Android with Talkback this doesn't help prevent these values being mixed in random order with everything else in the live region being announced as an update..

On Firefox Android with Talkback, when list item is appended/modified, all children of the list are announced again. Presumably the list is invalidated, it's not specific to lists, the ul and li elements can be swapped for divs, same behaviour. You'd need to make each their own live region and nest content with an inner div to avoid redundant announcements.

Other odd behaviour was experienced while iterating on this with other aria attribtues.

`aria-describedby` is used in the error items to provide additional context that it's the error label and not the form label when interacted with. Firefox ignores this however (unlike Chrome it announces list item context, although it's indexing/counting is broken making it less helpful).