Refactor inputs
Not entirely sure if this is beneficial. Extracted each input to it's own component, with an additional Expiry Date component using component slots to add the month and year.

None of these are really re-usable, so the only benefit is from further isolating their scopes and separating their layout order in the form?