Refactor `isNotExpired` validation to use generic superRefine type
Dropping old comments with variations on approach in favor of the generic `SuperRefinement<T>` type alias.

Additionally noting that the schema has differing input and output types. I haven't exported these yet, and have included a union type of the two as `zIO`(maybe not useful in practice?).

There can be valid reasons for wanting to have access to the zod object schema before the post/compound validation via the `superRefine()` for using with other schemas or accessing the `z.object()` methods external from this schema.