Use a context hook to pass through error state instead of prop drilling
Seems a little glued to the useForm logic.. Although the two lines of context could be extracted out to their own minimal file if that'd be useful..?

I haven't bothered to define a type for the context atm.

Using context to pass the error status to each input and error item seems like a good approach to avoid the repetitive prop drilling, especially for the inputs that I extracted out to one off components.