Provide schema as param to `useForm<T>`, resolving the typing issues
It seems the issues faced earlier was due to always using my specific schema that was imported, which technically prevented a generic type alias...

By passing it in as a parameter I can ensure the schema and it's output type match. This enables `T` to be passed to ZodError and remove the `| any` type union on `_onValid` data param :)

This submission hook logic is no longer tied to this component and can later be extracted out of it as a generic method/hook.

As such the two complimentary onValid/onInvalid methods that were exported have migrated back into the form component.