fix: Provide schema as param to `useForm<T>`, resolving the typing issues (#7)
It seems the type issues faced earlier were due importing my specific schema, which technically prevented a generic type alias... for some reason that didn't occur to me at the time :sweat_smile: 

By passing it in as a parameter I can ensure the schema and it's `<output>` (_data returned after validation, not prior input type_) type match. This enables `T` to be passed to ZodError and remove the `| any` union type I had on the `_onValid` data arg :grinning: 

This submission hook logic is no longer tied to this component and can later be extracted out of it as a generic method/hook :tada: 

As such the two complimentary `onValid()` + `onInvalid()` methods that were exported have migrated back into the main form component that they're specific to.